requirements.txt
python-binance==1.0.19
python-telegram-bot==20.3
websockets==10.4
aiohttp==3.9.3
